refactor(test): 테스트 일부 패키지 이동
잘못된 패키지 이동
