+++ src/test/java/lotto/domain/ProfitTest.java
@@ -1,4 +1,4 @@
-package lotto.dto;
+package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
